backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg[binary]==3.2.4
psycopg-pool>=3.2.0
httpx==0.27.2
pyjwt[crypto]==2.9.0
openai==1.12.0
beautifulsoup4==4.12.3
lxml==5.1.0
feedparser==6.0.11
pgvector==0.3.6
trafilatura>=2.0.0
tavily-python>=0.5.0
google-genai>=1.0.0
Pillow>=10.0.0
